fix: change checking for status value
The value for ServiceStatus should be an object instead returned null. In the Docker API documentation says that exists a parameter to include the status object in the response, but the .Net Package does not include this parameter.
It was decided to use the approach to obtain the status using the Replicas field.